requirements.txt
pandas
numpy
matplotlib
pyyaml
requests
yfinance
python-dateutil
pytz
